Allow local failures under dynamic execution to be ignored for exit codes > 127 (signals)
A local execution failing due to a signal is a sign of something outside of the actual action failing, for instance hitting `rlimits`.

Theoretically, this is not precise, as it relies on the return code indicating what signal killed the process. Tools could give such a return code on "normal" failures, but if it doesn't happen too often, that's not a big deal - it just means we would very occasionally fall back on remote. Tools could also catch the signal and turn it into a non-signal failure, which is not nice and should be reported as a bug. For workers, this is less of an issue, as we only see the return code when the worker actually dies, not the one returned for each action.

PiperOrigin-RevId: 493275597
Change-Id: Idd59858bd18d08e0e7cc1b6094e028d7ae067379