Add flag to skip printing specific aspects in build results
We have an internal system which tacks an aspect onto all the builds it runs,
but no one really cares about the aspect's outputs, just that with the default
--show_result their normal results are suppressed. This lets us avoid that.

There are a few ways we could have implemented this, I decided to go with the
simplest - non-repeated flag with comma-separated values. Practically, we could
evolve this to allow multiple and do something like we do with --output_group
with adding/subtracting from the existing/default, but we can cross that bridge
when/if we get there.

PiperOrigin-RevId: 603229525
Change-Id: I8b1a387eeb7e67cceaf4f8e821bd29c12b136982